+++ src/test/java/subway/persistance/LineDaoTest.java
@@ -1,4 +1,4 @@
-package subway.dao;
+package subway.persistance;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
