+++ src/test/java/subway/persistance/LineRepositoryImplTest.java
@@ -1,4 +1,4 @@
-package subway.dao;
+package subway.persistance;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -6,11 +6,10 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.test.annotation.Rollback;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.jdbc.Sql;
 import subway.Fixture;
-import subway.dao.entity.SectionEntity;
+import subway.persistance.entity.SectionEntity;
 import subway.domain.Line;
 import subway.domain.LineRepository;
