+++ src/test/java/subway/persistance/SectionDaoTest.java
@@ -1,4 +1,4 @@
-package subway.dao;
+package subway.persistance;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.jdbc.Sql;
-import subway.dao.entity.SectionEntity;
+import subway.persistance.entity.SectionEntity;
 import subway.domain.*;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
