+++ src/test/java/subway/persistance/StationDaoTest.java
@@ -1,4 +1,4 @@
-package subway.dao;
+package subway.persistance;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
